Struct nightingale_client::NightingaleClient

pub struct NightingaleClient { /* private fields */ }

Client that handles a single connection to a nightingale server.

Implementations

impl NightingaleClient

pub fn new_serenity( config: Config, handler: impl EventHandler + 'static ) -> Self

Creates a new instance to be used with serenity.

pub fn voice_manager(&self) -> Arc<dyn VoiceGatewayManager>

Returns a voice manager to be used with ClientBuilder#event_handler_arc

pub fn rest(&self) -> &RestClient

pub async fn connect(&mut self) -> Result<(), Error>

Connects to the server.

pub async fn disconnect(&mut self)

Disconnects from the server.

pub async fn reconnect(&mut self) -> Result<(), Error>

Reconnects to the server.

pub async fn join<G, C>(&self, guild: G, channel: C) -> Result<(), HttpError>where G: Into<NonZeroU64>, C: Into<NonZeroU64>,

Joins the given voice channel.

pub async fn leave<G: Into<NonZeroU64>>( &self, guild: G ) -> Result<(), HttpError>

Leaves the given voice channel.

pub async fn search<S>( &self, query: String, source: S ) -> Result<Vec<S::Track>, HttpError>where S: SearchSource,

Makes a search on the provided source.

pub async fn playlist<S>( &self, playlist: String, source: S ) -> Result<S::Playlist, HttpError>where S: SearchSource,

Gets the playlist items from the specified source.

pub fn get_player( &self, guild: impl Into<NonZeroU64> ) -> Option<Reference<'_, Player>>

Returns a reference to the player of the provided guild, if present.

pub fn get_player_mut( &self, guild: impl Into<NonZeroU64> ) -> Option<ReferenceMut<'_, Player>>

Returns a mutable reference to the player of the provided guild, if present.